+++ Backpack/Button/Classes/BPKButtonAppearanceSets.m
@@ -31,7 +31,7 @@
 
 #import <Backpack/Color.h>
 #import <Backpack/DarkMode.h>
-
+#import "UIColor+BPKButton.h"
 NS_ASSUME_NONNULL_BEGIN
 
 @implementation BPKButtonAppearanceSets
