+++ Backpack/Button/Classes/BPKButtonExtensions.swift
@@ -7,39 +7,6 @@
 
 import Foundation
 
-extension BPKButtonAppearance {
-    var hasGradientColor: Bool {
-        let hasStart = self.gradientStartColor != nil
-        let hasEnd = self.gradientEndColor != nil
-        return hasStart && hasEnd
-    }
-    
-    func gradient(traitCollection: UITraitCollection) -> BPKGradient? {
-        guard hasGradientColor else { return nil }
-        return gradientWith(
-            topColor: gradientStartColor!,
-            bottomColor: gradientEndColor!,
-            traitCollection: traitCollection
-        )
-    }
-    
-    private func gradientWith(
-        topColor top: UIColor,
-        bottomColor bottom: UIColor,
-        traitCollection: UITraitCollection
-    ) -> BPKGradient {
-        let direction = BPKGradientDirection.down
-        return BPKGradient(
-            colors: [
-                top.resolvedColor(with: traitCollection),
-                bottom.resolvedColor(with: traitCollection)
-            ],
-            start: BPKGradient.startPoint(for: direction),
-            end: BPKGradient.endPointFor(for: direction)
-        )
-    }
-}
-
 extension BPKButton {
     func themedAppearance(_ appearance: BPKButtonAppearance, highlighted: Bool) -> BPKButtonAppearance {
         var foreground: UIColor?
@@ -68,7 +35,7 @@ extension BPKButton {
             gradientStart = featuredGradientStartColor?.highlighted(highlighted)
             gradientEnd = featuredGradientEndColor?.highlighted(highlighted)
         case .link:
-            foreground = highlighted ? .reduceOpacity(of: linkContentColor) : linkContentColor
+            foreground = highlighted ? linkContentColor.reducedOpacity : linkContentColor
         default:
             break
         }
