+++ Backpack/Button/Classes/BPKButtonImagePosition.h
@@ -1,0 +1,5 @@
+typedef NS_ENUM(NSUInteger, BPKButtonImagePosition) {
+    BPKButtonImagePositionLeading = 0,
+    
+    BPKButtonImagePositionTrailing
+};
