--- Backpack/Button/Classes/BPKButtonImagePosition.swift
@@ -1,14 +1,0 @@
-//
-//  BPKButtonImagePosition.swift
-//  Backpack
-//
-//  Created by Nicolas Frugoni on 17/02/2022.
-//
-
-import Foundation
-
-@objc
-public enum BPKButtonImagePosition: NSInteger {
-    case leading
-    case trailing
-}
