+++ Backpack/Button/Classes/BPKButtonSize.h
@@ -1,0 +1,12 @@
+//
+//  BPKButtonStyle.h
+//  Pods
+//
+//  Created by Nicolas Frugoni on 18/02/2022.
+//
+
+typedef NS_ENUM(NSUInteger, BPKButtonSize) {
+    BPKButtonSizeDefault = 0,
+    
+    BPKButtonSizeLarge = 1
+};
