+++ Backpack/Dialog/Classes/BPKDialogContentView.m
@@ -22,7 +22,7 @@
 #import <Backpack/Common.h>
 #import <Backpack/Label.h>
 #import <Backpack/Spacing.h>
-#import <Backpack/Backpack-Swift.h>
+#import <Backpack/Button.h>
 
 NS_ASSUME_NONNULL_BEGIN
 @interface BPKActionButtonPair : NSObject
