+++ Backpack/NavigationBar/Classes/BPKNavigationBar.m
@@ -22,7 +22,7 @@
 #import <Backpack/DarkMode.h>
 #import <Backpack/Label.h>
 #import <Backpack/Spacing.h>
-
+#import "BPKButtonImagePosition.h"
 #import "BPKNavigationBarLargeTitleView.h"
 #import "BPKNavigationBarTitleView.h"
 #import <Backpack/Backpack-Swift.h>
