+++ Backpack/Nudger/Classes/BPKNudger.m
@@ -18,7 +18,7 @@
 
 #import "BPKNudger.h"
 
-#import <Backpack/Backpack-Swift.h>
+#import <Backpack/Button.h>
 #import <Backpack/Common.h>
 #import <Backpack/Icon.h>
 #import <Backpack/Label.h>
