+++ Backpack/Snackbar/Classes/BPKSnackbar.m
@@ -18,12 +18,14 @@
 
 #import "BPKSnackbar.h"
 
+
 #import "BPKSnackbarButton.h"
 
 #import <Backpack/Color.h>
 #import <Backpack/Duration.h>
 #import <Backpack/Label.h>
 #import <Backpack/Spacing.h>
+#import "BPKButtonSize.h"
 #import <Backpack/Backpack-Swift.h>
 
 NSString *const SnackbarAccessibilityIdentifier = @"snackbarView";
