+++ Backpack/Theme/Classes/BPKDefaultTheme.m
@@ -17,7 +17,7 @@
  */
 
 #import "BPKDefaultTheme.h"
-#import <Backpack/Backpack-Swift.h>
+
 #import <Backpack/Chip.h>
 #import <Backpack/Color.h>
 #import <Backpack/Font.h>
