+++ Backpack/Theme/Classes/BPKLondonTheme.m
@@ -17,7 +17,6 @@
  */
 
 #import "BPKLondonTheme.h"
-#import <Backpack/Backpack-Swift.h>
 #import <Backpack/Chip.h>
 #import <Backpack/Color.h>
 #import <Backpack/Font.h>
